// backend/src/modules/production-orders/dto/create-production-order-post-preparation-solution-check.dto.ts
export class CreateProductionOrderPostPreparationSolutionCheckDto {
  solution_color?: string | null;
  solution_clarity?: string | null;
  solution_ph_1?: number | string | null;
  solution_ph_2?: number | string | null;
  solution_ph_3?: number | string | null;
  checked_by_id?: number | string | null;
}